--- client/src/components/common/SubmitButton.vue
@@ -1,99 +1,0 @@
-<template>
-  <button
-    type="submit"
-    class="btn_submit"
-    :class="{ loading }"
-    :disabled="disabled || loading"
-    @click="submitHandler"
-  >
-    {{ text }}
-  </button>
-</template>
-
-<script>
-export default {
-  name: "SubmitButton",
-  props: {
-    text: String,
-    disabled: Boolean,
-    loading: Boolean,
-    submitHandler: Function,
-  },
-};
-</script>
-
-<style scoped>
-@charset "utf-8";
-
-.btn_submit {
-  position: relative;
-  overflow: hidden;
-  width: 100%;
-  border: 0;
-  border-radius: 19px;
-  background-color: #1da1f2;
-  color: #fff;
-  font-size: 1.5rem;
-  font-weight: 700;
-  line-height: 38px;
-  text-align: center;
-  cursor: pointer;
-}
-
-.btn_submit.loading:before {
-  content: "";
-  position: absolute;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  background-color: #1da1f2;
-}
-
-.btn_submit.loading:after {
-  content: "";
-  position: absolute;
-  width: 16px;
-  height: 16px;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  margin: auto;
-  border: 4px solid transparent;
-  border-top: 4px solid #fff;
-  border-radius: 50%;
-  animation: button-loading-spinner 1s ease infinite;
-}
-
-@keyframes button-loading-spinner {
-  from {
-    transform: rotate(0turn);
-  }
-
-  to {
-    transform: rotate(1turn);
-  }
-}
-
-.btn_submit[disabled] {
-  opacity: 0.5;
-  cursor: auto;
-}
-
-.btn_submit:hover {
-  background-color: #1a91da;
-  transition: background-color 0.2s;
-}
-
-.btn_submit[disabled]:hover {
-  background-color: #1da1f2;
-}
-
-.btn_submit:focus {
-  background-color: #1a91da;
-  box-shadow: #8ed0f9 0 0 0 2px;
-  outline: none;
-  transition: background-color 0.2s box-shadow 0.2s;
-}
-</style>
